+++ EventMessageVisualisationStreamer/src/main/java/Image/FrameImage.java
@@ -45,7 +45,7 @@ public class FrameImage implements ImageInterface {
         if (!image.containsKey(detector)) {
             throw new InvalidParameterException(MISSING_KEY_ERROR_MESSAGE);
         }
-        return (int) image.get(detector);
+        return (long) image.get(detector);
     }
 
     public void setFrequency(long detector, long newFreq) {
@@ -58,9 +58,9 @@ public class FrameImage implements ImageInterface {
     }
 
     public void incrementFrequency(long detector) {
-        int oldFreq = 0;
+        long oldFreq = 0L;
         if (image.containsKey(detector)) {
-            oldFreq = (int) image.get(detector);
+            oldFreq = (long) image.get(detector);
         } else if (detector < 0) {
             throw  new InvalidParameterException(DETECTOR_ID_POSITIVE_ERROR_MESSAGE);
         }
